requirements.txt
psycopg2_binary==2.9.10
python-dotenv==1.1.0
Requests==2.32.3
